--- a/java-frontend/src/main/java/org/sonar/java/JavaTestClasspath.java
+++ b/java-frontend/src/main/java/org/sonar/java/JavaTestClasspath.java
@@ -19,7 +19,6 @@
  */
 package org.sonar.java;
 
-import com.google.common.collect.Lists;
 import org.sonar.api.batch.fs.FileSystem;
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.config.Settings;
@@ -28,7 +27,8 @@ import org.sonar.api.utils.log.Loggers;
 import org.sonar.api.utils.log.Profiler;
 
 import java.io.File;
-import java.util.List;
+import java.util.ArrayList;
+import java.util.Set;
 
 public class JavaTestClasspath extends AbstractJavaClasspath {
 
@@ -44,13 +44,13 @@ public class JavaTestClasspath extends AbstractJavaClasspath {
       validateLibraries = fs.hasFiles(fs.predicates().all());
       Profiler profiler = Profiler.create(LOG).startInfo("JavaTestClasspath initialization");
       initialized = true;
-      binaries = getFilesFromProperty(JavaClasspathProperties.SONAR_JAVA_TEST_BINARIES);
-      List<File> libraries = getFilesFromProperty(JavaClasspathProperties.SONAR_JAVA_TEST_LIBRARIES);
-      if(libraries.isEmpty()) {
+      binaries = new ArrayList<>(getFilesFromProperty(JavaClasspathProperties.SONAR_JAVA_TEST_BINARIES));
+      Set<File> libraries = getFilesFromProperty(JavaClasspathProperties.SONAR_JAVA_TEST_LIBRARIES);
+      if (libraries.isEmpty()) {
         LOG.warn("Bytecode of dependencies was not provided for analysis of test files, you might end up with less precise results. " +
-            "Bytecode can be provided using sonar.java.test.libraries property");
+          "Bytecode can be provided using sonar.java.test.libraries property");
       }
-      elements = Lists.newArrayList(binaries);
+      elements = new ArrayList<>(binaries);
       elements.addAll(libraries);
       profiler.stopInfo();
     }