… java-frontend/src/test/files/classpath/lib/.gitignore
… java-frontend/src/test/files/classpath/lib/target/classes/foo.jar