The goal is to optimize the initialization of main and test classpaths, and also to get rid of old utilities to find files recursively and switch to JDK 7 NIO, with a much more readable code.
Here are the benchmark results when doing the following searches in my local mvn repo, with **4903 jars**:

- /repository
**`952ms -> 85ms`**
- /repository/
`735ms -> 84ms`
- /repository/* * /*.jar
`321ms -> 175ms`
- /repository/* *.jar
`291ms -> 168ms`
- /repository/* * /org/**.jar
`342ms -> 169ms`
- /repository/* * /* 
**`3160ms -> 220ms`**
- list of all 4903 files
`10ms -> 10ms`

Note that the syntax for patterns is now strictly GLOB, meaning that Windows path separators won't work, at least on Linux (one test removed for this).