@dbmeneses 2016-08-11
@benzonico I changed this to:

- Adapt path separator to the platform (test restored)
- Calculate internally a Set to avoid dups, but exposes in the end a List because I don't want to change all the classes using the classpath (can be done later). Performance hit in the worst case tested was from `220ms -> 231ms`.