@dbmeneses 2016-08-15
Added the special case for target/classes and target/test-classes with a unit test.